+++ docs/CHANGELOG.rst
@@ -21,6 +21,8 @@ Changed
   inputs in ``upload-gtf`` and ``upload-gff3`` process
 * **BACKWARD INCOMPATIBLE:** Track ``species`` and ``build``
   information in aligner processes
+* **BACKWARD INCOMPATIBLE:** Require ``species`` and ``build``
+  inputs in ``upload-bam`` and ``upload-bam-indexed`` process
 
 Added
 -----